+++ library/src/main/java/com/mlsdev/livedatasocialauth/library/common/Account.kt
@@ -1,5 +1,7 @@
 package com.mlsdev.livedatasocialauth.library.common
 
+import android.net.Uri
+
 class Account {
     var id: String? = null
     var displayName: String? = null
@@ -7,6 +9,7 @@ class Account {
     var lastName: String? = null
     var email: String? = null
     var authProvider: AuthProvider? = null
+    var avatar: Uri? = null
 }
 
 enum class AuthProvider(val value: String) {